redirection to book detail
