+++ react/wbooks/src/app/components/BookInfo/index.js
@@ -1,28 +1,29 @@
-import React from 'react';
-import { string } from 'prop-types';
+import React, { Component } from 'react';
+import { withRouter } from 'react-router-dom';
 
-import imagePlaceholder from '../../assets/book-cover.png';
+import { Routes } from '../../../constants';
 
-import styles from './styles.module.scss';
+import BookInfo from './layout';
 
-function BookInfo({ title, author, className }) {
-  return (
-    <div className={`column ${styles.container} ${className}`}>
-      <img src={imagePlaceholder} alt="Wolox book cover" className={styles.woloxBookCoverImage} />
-      <h2 className={`${styles.title} m-top-2`}>{title}</h2>
-      <span className={`${styles.author} m-top-2`}>{author}</span>
-    </div>
-  );
+class BookInfoContainer extends Component {
+  handleOnClick = () => {
+    const { title, author, id } = this.props;
+    this.props.history.push({
+      pathname: Routes.BOOK_DETAIL.replace(':id', id),
+      state: {
+        title,
+        author,
+        genre: 'magia'
+      }
+    });
+  };
+
+  render() {
+    const { title, author, className } = this.props;
+    return (
+      <BookInfo title={title} author={author} className={className} handleOnClick={this.handleOnClick} />
+    );
+  }
 }
 
-BookInfo.propTypes = {
-  author: string.isRequired,
-  title: string.isRequired,
-  className: string
-};
-
-BookInfo.defaultProps = {
-  className: ''
-};
-
-export default BookInfo;
+export default withRouter(BookInfoContainer);
