+++ react/wbooks/src/app/components/Routes/index.js
@@ -15,7 +15,7 @@ function AppRoutes() {
     <Router>
       <Switch>
         <AuthenticatedRoute isPublicRoute path={Routes.SIGN_UP} component={SignUp} />
-        <AuthenticatedRoute isPublicRoute path="/hola" component={BookDetail} />
+        <AuthenticatedRoute isPublicRoute={false} path={Routes.BOOK_DETAIL} component={BookDetail} />
         <HybridRoute exact path={Routes.LOGIN_AND_HOME} privateComponent={Home} publicComponent={Login} />
       </Switch>
     </Router>
