+++ react/wbooks/src/constants.js
@@ -1,4 +1,5 @@
 export const Routes = {
   LOGIN_AND_HOME: '/',
-  SIGN_UP: '/sign_up'
+  SIGN_UP: '/sign_up',
+  BOOK_DETAIL: '/books/:id'
 };
